requirements.txt
flask==2.3.2
flask-cors==4.0.0
flask-pymongo==2.3.0
python-dotenv==1.0.0
requests==2.31.0
librosa==0.10.1
numpy==1.24.3
joblib==1.3.1
bson==0.5.10
soundfile==0.12.1